+++ src/errors/not-found.error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app.error';
+
+export class NotFoundError extends AppError {
+  constructor(message = 'Not Found.') {
+    super(message, 'NotFoundError');
+  }
+}